// packages/web/src/presentation/store/types.ts
export interface HACCPSlice {
    pccs: PCC[];
    haccpLogs: HACCPLog[];
    haccpTasks: HACCPTask[];
    haccpTaskCompletions: HACCPTaskCompletion[];
    haccpTimers: HACCPTimer[];
    setPCCs: (pccs: PCC[]) => void;
    setHACCPLogs: (logs: HACCPLog[]) => void;
    setHACCPTasks: (tasks: HACCPTask[]) => void;
    setHACCPTaskCompletions: (completions: HACCPTaskCompletion[]) => void;
    setHACCPTimers: (timers: HACCPTimer[]) => void;
    addPCC: (pcc: PCC) => void;
    updatePCC: (pcc: PCC) => void;
    deletePCC: (id: string) => void;
    addHACCPLog: (log: HACCPLog) => void;
    addHACCPTask: (task: HACCPTask) => void;
    updateHACCPTask: (task: HACCPTask) => void;
    deleteHACCPTask: (id: string) => void;
    completeHACCPTask: (completion: HACCPTaskCompletion) => void;
    addHACCPTimer: (timer: HACCPTimer) => void;
    updateHACCPTimer: (timer: HACCPTimer) => void;
    deleteHACCPTimer: (id: string) => void;
}
// ...
export interface AppState extends